tanzeel-ai/server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.14.0
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.18
httpx>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
